fix(condo): DOMA-6670 fixed default dropdown
